requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses==0.6
Django==4.2
django-dotenv==1.4.2
django-environ==0.11.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
idna==3.7
payze-pkg==2.4
pydantic==2.4.2
pydantic_core==2.10.1
python-click==0.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
